+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - Modern JavaScript From The Beginning (Udemy)
   - Section 2: JavaScript Language Fundamentals
     - General Loops
+    - A Look At The Window Object
 - An example with loops for General Loops was added.
+- An example with the window object for A Look At The Window Object was added.
 
 ## August 5, 2020
 - Modern JavaScript From The Beginning (Udemy)